+++ tests/test.nim
@@ -24,6 +24,7 @@ import ./test_dwcas_pair_ptr
 import ./test_dwcas_pair_alignment
 import ./test_dwcas_pair_shape_positive
 import ./test_dwcas_memory_orders
+import ./test_dwcas_fetch_ops
 import ./t_atomics_dsl
 import ./t_thread_id
 import ./t_item_processing